[RFC] Remove utilities from core rebound
This removes a number of the core utilities from rebound in an effort to
reduce its size. This brings it down from ~3.95kb gzip to ~3.5kb gzip.

These utilities, which do things like interpolate colors, aren't part of
the core purpose of rebound, but are often used with it. Continue to
make these available by `require('rebound/dist/ReboundMathUtil')`, etc.
