+++ package.json
@@ -30,7 +30,7 @@
     "jest": "jest",
     "watch": "yarn build -- --watch",
     "bundlesize": "NODE_ENV=production yarn build && bundlesize",
-    "test": "yarn lint && yarn flow && yarn jest && yarn bundlesize"
+    "test": "yarn lint && yarn flow check && yarn jest && yarn bundlesize"
   },
   "devDependencies": {
     "babel-core": "^6.22.0",
